data_pipeline/requirements.txt
requests==2.31.0
pandas==2.0.1
python-dotenv==1.0.0
sentinelsat==1.1.1
geopandas==0.13.0
shapely==2.0.1
sqlalchemy==2.0.12
geoalchemy2==0.13.0
psycopg2-binary==2.9.6
pystac-client==0.6.1
rasterio==1.3.6
netCDF4==1.6.3
xarray==2023.4.2
apache-airflow==2.6.1
pydantic==1.10.7
planetary-computer==0.4.3
tqdm==4.65.0
boto3==1.26.137
tenacity==8.2.2
